backend/requirements.txt
azure-functions
torch
torchvision
Pillow
numpy
azure-storage-blob
timm
